add value changed signal for custom interpolation parameter wrapper move duplicated code into common file
